Shell script to ease scans from the command-line.
Performs copyright scan on each of the command-line arguments.

If the argument is a directory, scans all of the files contained in the
directory or in subdirectories.

If the argument is a file, scans the file.

The script tries to see through symbolic links to the actual
destination path.

Use the --deep flag to extract and scan files embedded in archives.

Or the --nodeep to turn off deep scanning for subsequent command-line
arguments.

Change-Id: Iba04ead63c876f6e6020cfc40d2933002a3ee70f