Fix shell script directories.
The build flags affect the `bazel info bazel-bin` output and must match
those used during `bazel build`.

The path to the files from the data attribute of the build rule differs
when build from gerrit than when built from plugins/copyright, but the
differences are reflected in the ${TEST_BINARY} dirname.

Change-Id: Ide5bb8abdb39c5835b3b9a00b1f67abf2a0e7e46