CopyrightConfig to manage plugin configurations.
Copyright plugin part 2 of 3: CopyrightConfig manages and validates
plugin configurations described in the gerrit.config file, and in the
All-Projects project.config and groups files.

ScannerConfig wraps all of the parameters needed for scanning files.

CheckConfig wraps facilitates validation of the configuration. It also
provides the main entry point for a command-line tool to check for
performance-harming regular expression patterns.

When a new configuration changes the scanner pattern, the gerrit server
may be configured to require the submitter run the command-line tool
and copy its output to the commit message. By default, the server will
require this process, which can be turned off by making the timeTestMax
value 0.

Change-Id: I82c034be01b9b33b1f57041627d6065e27b021db