CopyrightValidator to scan uploaded files.
Copyright plugin part 3 of 3: CopyrightValidator scans each new
revision looking for copyright author/owner or license declarattions.

Whenever it finds a copyright declaration that might not be allowed, it
requires special review by voting down a configured label and by adding
a configured reviewer.

e.g. If it finds a third-party license or owner outside of the projects
where third-party code is allowed, it requires special review.

If it finds a forbidden or unknown license anywhere, it requires
special review.

If it finds a first-party license--including first-party licensed code
from a third-party, it notes the finings but requires no special
review.

etc.

Change-Id: Ia272fd02b3f38c5649dfc9fd8cbe3521cdfa4efc