Play better with automatic rotations.
If a rotation alias is added to REVIEWER, the rotation feature will
move it to CC. Moving it back to REVIEWER could trigger adding
another reviewer for each revision. Don't add a reviwer if already
in reviewers list.

Upgraded CopyrightPatternsTest to the latest version of Truth to fix
broken test.

Change-Id: Ifd31baad8e9031c1d4b50ad766f93d2c99349691